lots of cleanup clock stuff bug fixes client capabilities! object store attributes, collections (untested in FakeStore)
git-svn-id: https://ceph.svn.sf.net/svnroot/ceph@463 29311d96-e01e-0410-9327-a35deaab8ce9

+++ ceph/client/Buffercache.cc
@@ -1,6 +1,6 @@
 #include "Buffercache.h"
 
-#include "include/config.h"
+#include "config.h"
 #undef dout
 #define  dout(l)    if (l<=g_conf.debug) cout << "client" << "." << pthread_self() << " "
 
@@ -51,7 +51,7 @@ void Bufferhead::alloc_buffers(size_t size)
 {
   dout(10) << "bc: allocating buffers size: " << size << endl;
   while (size > 0) {
-    if (size <= g_conf.client_bcache_alloc_maxsize) {
+    if (size <= (unsigned)g_conf.client_bcache_alloc_maxsize) {
           size_t k = g_conf.client_bcache_alloc_minsize;
           size_t asize = size - size % k + (size % k > 0) * k;
 	  buffer *b = new buffer(asize);
@@ -165,7 +165,7 @@ void Bufferhead::claim_append(Bufferhead *other)
 void Dirtybuffers::erase(Bufferhead* bh) 
 {
   dout(7) << "dirtybuffer: erase bh->ino: " << bh->ino << " offset: " << bh->offset << endl;
-  int osize = _dbufs.size();
+  unsigned osize = _dbufs.size();
   for (multimap<time_t, Bufferhead*>::iterator it = _dbufs.lower_bound(bh->dirty_since);
        it != _dbufs.upper_bound(bh->dirty_since);
        it++) {
